fix(Facade): remove isPlaying check from change handler
There was an issue with Malimbe where change handler properties
were running at edit time which would cause issues with the setup
of the prefab. The original solution was to add in an isPlaying
check to prevent this. A Malimbe fix has now been pushed so the
latest version of Zinnia now contains the fix in Malimbe and therefore
this check is now no longer required.