feat(Facade): expose device details and provide new events
The DeviceDetailsRecord which can now be stored on a CameraRig is now
exposed via the TrackedAliasFacade so it's possible to get the current
CameraRig's device details as well as the current determined dominant
controller.

There are also new events that have been added to the Facade that
bubble up the events from the internal device records to make it
easier to know when device statuses change.