feat(prefab): provide dominant controller alias
The dominant controller alias is now part of the TrackedAlias Aliases
collection and will track the position/orientation of whatever the
current dominant controller is.

This can be used to set up tracking objects to whatever the dominant
controller is rather than having to always use the left/right
controller.

There are also new events that are raised when the left/right
controller become dominant or when the headset becomes the dominant
controller.

Additional Facade properties for dominant controller data has also
been added for completeness.